requirements.txt
streamlit
streamlit_option_menu
requests
streamlit_lottie
Pillow

